bugfix in initialization of modified binary mera network some functions moved to seperate file modified_binary_mera_lib.py
